port the API to Vercel serverless functions
