+++ package.json
@@ -1,26 +1,13 @@
 {
   "name": "token-supply",
   "license": "GPL-3.0",
-  "main": "index.js",
-  "scripts": {
-    "dev": "concurrently \"npx tsc --watch\" \"nodemon -q dist/index.js\"",
-    "build": "npx tsc",
-    "start": "node dist/index.js"
-  },
   "dependencies": {
-    "cors": "^2.8.5",
     "dotenv": "^16.0.1",
-    "ethers": "^5.6.9",
-    "express": "^4.18.1",
-    "express-rate-limit": "^6.5.1",
-    "helmet": "^5.1.1"
+    "ethers": "^5.6.0"
   },
   "devDependencies": {
-    "typescript": "^4.7.4",
-    "@types/cors": "^2.8.12",
-    "@types/express": "^4.17.13",
-    "@types/node": "^18.6.4",
-    "concurrently": "^7.3.0",
-    "prettier": "^2.7.1"
+    "@vercel/node": "^2.5.7",
+    "prettier": "^2.7.1",
+    "typescript": "^4.7.4"
   }
 }
